+++ cartography/intel/okta/applications.py
@@ -5,8 +5,8 @@ import logging
 from okta import AppInstanceClient
 from okta.framework.OktaError import OktaError
 
-from cartography.intel.okta import create_api_client
-from cartography.intel.okta import is_last_page
+from cartography.intel.okta.utils import create_api_client
+from cartography.intel.okta.utils import is_last_page
 
 logger = logging.getLogger(__name__)
